Added item leveling (not completed yet) Fixed dojo exit Fixed some bugs
git-svn-id: https://svn6.assembla.com/svn/MoopleDEV@100 59296cf3-ab24-4890-a14e-c0dede3ece5d

+++ src/client/IEquip.java
@@ -48,7 +48,7 @@ public interface IEquip extends IItem {
     public short getJump();
     public short getVicious();
     public int getItemExp();
-    public int getItemLevel();
+    public byte getItemLevel();
     public int getRingId();
     public void setRingId(int id);
     public boolean isWearing();
